[api] Put some order in API files.
We start to sort out the dependency mess of API by providing a logical
declaration order in `API.{ml,mli}` that is related to the actual
declaration order in Coq.

In particular, we solve problems such as `Summary.frozen` being used
in `API.mli` but not declared by the `Summary` exported module. Many
more problems like these existed, as in general, in the previous order
the likelihood for a type to refer to a local or a Coq module was
basically random.

Unfortunately this problem persists to some extent; for example
`Globnames` is referenced from intf, however it cames after it in the
linking order.

OCaml upstream has warned repeatedly that it is not OK to reference
objects in this way.

Due to the correct declaration order, we can remove now the spurious
module `API.Prelude` as well as some spurious declarations present
there.

A few definitions had to be moved out of `intf` due to them depending
on `EConstr` now.

Unfortunately we don't solve the main problems with `ml` files yet,
however this should be orthogonal to that problem. It is becoming
clear that OCaml doesn't really support well having "`mli`-only"
files.

We also had to move a bunch of modules (`Impargs`, `Declare`,
`Ind_tables`, `Miscprint`) to their proper place as they were linked
differently, for instance as they now depend on `EConstr` or the "new"
proof engine depends on them.