[plugins] Support for a findlib-based loader.
We introduce new syntax
```
Require ML Module pkg1 ... pkgn.
```
that uses findlib to locate and load a Coq plugin and its dependencies.

This makes feasible to actually develop Coq plugins that depend on
OCaml libraries not linked into the Coq main binaries without hitting
problems.

Fixes #5028, fixes #7698 .

This provides a workaround for #12607 , and helps with #9547 .