[build] [tools] Consolidation of directory handling code
This amounts to moving some of the bits on `Envars` to a new library
`coq.boot`, which is lighter so tools don't have to pull the whole of
`clib` in order to use this.

We take advantage of the move to introduce a stronger API, which
introduces the notion of "Coq runtime enviroment", which can be later
used to access different elements of a Coq install.

In addition to the removal of duplicate and inconsistent code, this
API should make a few things easier, including improvements
w.r.t. re-locatable setups.

We also remove some remnants of the coq-core compatibility layer,
which is not needed after the switch to Dune for the OCaml parts of
Coq.