[build] Bump to dune-lang 2.8
We bump the build language version so we can use the `(env (coq ...))`
feature for flags.

This was introduced in Dune 2.8, however Dune 2.9 is required due to a
bug. Now Dune adds the -q and native compiler flag, so they can be
dropped from the stdlib's flags.

We also have to tell dune we want absolute control over the C flags
for the VM.