[configure] Remove unused variables, fix prefix detection
[This PR is still not ready, but necessary to fix some issues with
`make install` on 8.14, submitting to signal intent to work on this]

Since the Dune transition, Coq delegates a large part of the install
to Dune itself; this eliminates duplication of configuration concerns,
as now the location of binaries for example is set in a central
place. Prefix is however still in control of the user.

We cherry pick from ongoing PRs to fix some discrepancies on
configuration variables in configure for 8.14:

- `COQBIN` and `coqdocdir` are not used anymore
- `COQPREFIX` must be set in the interactive procedure

TODO: what about profile devel + install?

Fixes #14468