[vernacstate] Remove `Vernacstate.Parser.t`
We remove `Vernacstate.Parser.t` as it is just a duplicate of
`Vernacstate.Synterp`.

The module was originally introduced to serve the same purpose.

However things still don't work in the stm if we unfreeze the whole
synterp state instead of just the parsing state.

I'd guess this is an stm problem, likely due freezing the synterp
state at the wrong time? (just after parsing, but before synterp has
been run?) It is hard to tell without more research.

I thus wonder how we should proceed here, for now I kept the original
STM semantics.

Follow up to #19187